cleanup color usage and add initial dark mode
Makes use of the relatively new colorScheme from the Theme class and
switches use from the old ThemData special colors and most of the
many hard coded ones we had sprinkled all over the place to that one.
Note, this is not finished and it can still be improved a bit, but
basic theming should be quite ok and the remaining improvements can
be done way easier gradually, with the feedback we get once this is
out, as else I think this would be never finished..

Note: This patch does quite a bit more than that, and I'm not to
happy to have all the null-safety, unrelated refactoring and bug
fixes squashed into this commit but the color cleanup was a mammoth
task and touched so much that I also saw so much to cleanup.

While I picked the low-and-easy fruits of untangling and committed
those up front, doing so for all of the changes now is just to
unfeasible, it would require me quite a few hours to do cleanly; so,
lets just ignore style rules for this one and go ahead..

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>