transform android splash screen logo to vector
The new Android 12 splash screen will mask the logo / icon.
> As with adaptive icons, one-third of the foreground is masked (3).
-- figure 2, [0]

This resulted in the current logo being cut off by the circular mask,
appearing broken.

[0] https://developer.android.com/guide/topics/ui/splash-screen

Create a new version, based on the stacked variant accounting for the
additional space around the logo needed for the mask and transform it
to a (android) vector drawable via the android studio's Asset Studio
import wizard, see the README.md for more details.

The source svg file plus a template should we need to change that in the
future, are stored in the assets directory.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>